requirements.txt
pymupdf>=1.24.0
requests>=2.31.0
anthropic>=0.34.0
Pillow>=10.0.0
